First commit
We pushed the main.py, the requirements.txt and the (modified) detoxify library. Install the detoxify library first with 'pip install -e detoxify' ; the requirement with 'pip install requirements.txt' and the run the main.py

We still have some bugs in main.py (vocal_answer function). Bark generated a maximum of 13 seconds of outputs, therefore we are segmenting the sentence in multiple inputs that we feed it. The bug is only about concatenating the audio arrays into one single array and converting it into an audio file. The patch will be pushed tonight.